Moved "clear nat" to separate module - nat.py
