+++ clear/main.py
@@ -10,7 +10,6 @@ from show.plugins.pbh import read_pbh_counters
 from config.plugins.pbh import serialize_pbh_counters
 from . import plugins
 
-
 # This is from the aliases example:
 # https://github.com/pallets/click/blob/57c6f09611fc47ca80db0bd010f05998b3c0aa95/examples/aliases/aliases.py
 class Config(object):
@@ -431,30 +430,10 @@ def line(target, devicename):
     click.echo(output)
     sys.exit(exitstatus)
 
-#
-# 'nat' group ("clear nat ...")
-#
-
-@cli.group(cls=AliasedGroup)
-def nat():
-    """Clear the nat info"""
-    pass
-
-# 'statistics' subcommand ("clear nat statistics")
-@nat.command()
-def statistics():
-    """ Clear all NAT statistics """
-
-    cmd = "natclear -s"
-    run_command(cmd)
-
-# 'translations' subcommand ("clear nat translations")
-@nat.command()
-def translations():
-    """ Clear all NAT translations """
-
-    cmd = "natclear -t"
-    run_command(cmd)
+#add nat commands ("clear nat statistics", "clear nat translations")
+from . import nat
+cli.add_command(nat.statistics)
+cli.add_command(nat.translations)
 
 # 'pbh' group ("clear pbh ...")
 @cli.group(cls=AliasedGroup)
